+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## Unreleased
 ### Changed
 * Python >=3.9 required
+* `DateTimeField.within` defaults to local time
 
 ### Fixed
 * Deprecation warnings